menus agregados: seleccionar piso y patron, empezando graficar

+++ Clases/Piso.py
@@ -1,6 +1,6 @@
 from Listas.ListaPatrones import ListaPatrones
 class Piso:
-    def __init__(self, nombre: str, filas: int, columnas: int, precio_voltear: int, precio_intercambiar: int, lista_patrones: ListaPatrones):
+    def __init__(self, nombre: str, filas: int, columnas: int, precio_voltear: float, precio_intercambiar: float, lista_patrones: ListaPatrones):
         self.nombre = nombre
         self.filas = filas
         self.columnas = columnas
